feat: improve loading spectra loading spectra or molecule files by passing an array of buffer refactor useLoadSpectra hook
