+++ src/events/types.ts
@@ -1,10 +1,14 @@
 import { NMRiumData } from 'nmrium';
 import { State } from 'nmrium/lib/component/reducer/Reducer';
+import { File } from '../hooks/useLoadSpectra';
 
-type EventType = 'load' | 'dataChange' | 'loadURLs' | 'error';
+type EventType = 'load' | 'dataChange' | 'loadURLs' | 'loadFiles' | 'error';
 
 interface LoadURLs {
   urls: string[];
+}
+interface LoadFiles {
+  files: File[];
 }
 
 type EventData<T extends EventType> = T extends 'load'
@@ -13,6 +17,8 @@ type EventData<T extends EventType> = T extends 'load'
   ? State
   : T extends 'loadURLs'
   ? LoadURLs
+  : T extends 'loadFiles'
+  ? LoadFiles
   : T extends 'error'
   ? Error
   : never;
