+++ src/utilities/castToPromise.ts
@@ -1,0 +1,7 @@
+export function castToPromise(func, ...props) {
+  return new Promise((resolve) => {
+    const spectra = [];
+    func(spectra, ...props);
+    resolve(spectra);
+  });
+}
