+++ src/utilities/getFileNameFromURL.ts
@@ -1,0 +1,3 @@
+export function getFileNameFromURL(url: string) {
+  return url.substring(url.lastIndexOf('/') + 1);
+}
